rotas concluídas

+++ app/Http/Controllers/api/ClientController.php
@@ -6,6 +6,7 @@ use App\Http\Controllers\Controller;
 use App\Http\Resources\Client as ResourcesClient;
 use App\Models\Client;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class ClientController extends Controller
 {
@@ -14,8 +15,9 @@ class ClientController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+    public function index(Request $request)
     {
+        Gate::authorize('viewAny', Client::class);
         return ResourcesClient::collection(Client::all());
     }
 
@@ -27,6 +29,7 @@ class ClientController extends Controller
      */
     public function store(Request $request)
     {
+        Gate::authorize('create', Client::class);
         $client = Client::create($request->all());
         return new ResourcesClient($client);
     }
@@ -39,6 +42,7 @@ class ClientController extends Controller
      */
     public function show(Client $client)
     {
+        Gate::authorize('view', $client);
         return new ResourcesClient($client);
     }
 
@@ -51,6 +55,7 @@ class ClientController extends Controller
      */
     public function update(Request $request, Client $client)
     {
+        Gate::authorize('update', $client);
         $client->update($request->all());
         return new ResourcesClient($client);
     }
@@ -63,9 +68,25 @@ class ClientController extends Controller
      */
     public function destroy(Client $client)
     {
+        Gate::authorize('delete', $client);
         $client->delete();
         return response()->json([
             'message' => 'Apagado com sucesso'
+        ]);
+    }
+
+    /**
+     * Remove the specified resource from storage.
+     *
+     * @param  \App\Models\Client  $client
+     * @return \Illuminate\Http\Response
+     */
+    public function destroyAll()
+    {
+        Gate::authorize('deleteAll', Client::class);
+        Client::truncate();
+        return response()->json([
+            'message' => 'Todos os registros foram excluídos'
         ]);
     }
 }
